\name{devx}
\alias{devx}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate Variance of an integer 
}
\description{
C code for calculating varinace of an integer x
}
\usage{
devx(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A vector of numeric Values 
}
}
\details{
Calculate the variance using C loop
}
\value{
Input integer vector
}
\author{
Md. Shahjaman
}


